# API_Design/YAML+Part5/LTMS_OpenAPI_MVP.yaml
openapi: 3.1.0
info:
  title: LTMS API — MVP (R01)
  version: "0.1.0"
  description: >
    Local Tournament Management System — เฉพาะ endpoint ที่อยู่ในเฟส MVP (92 endpoint)
    ครอบคลุมเส้นทางหลัก OF-01 ถึง OF-03 ตั้งแต่ขอจัดตั้งทัวร์นาเมนต์จนถึงประกาศผู้ชนะ
    ดูฉบับ Sprint #1/#2 ที่ LTMS_OpenAPI_Sprint1-2.yaml (แยกไฟล์ตามกำหนดส่งมอบใน SRS หัวข้อ 2.7)
  contact:
    name: ทีม WWW.
servers:
  - url: https://api.ltms.example.com/api/v1
    description: Production (ตัวอย่าง — เปลี่ยนตาม deployment จริง)
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token อายุ 7 วัน (ตั้งค่าผ่าน JWT_EXPIRES_IN) — ไม่มี refresh token
        (ดูเหตุผลใน LTMS_API_Design_Part0-1.md หัวข้อ 1.3)
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    TeamRef:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        sportTypeId: { type: integer }
      required: [id, name, sportTypeId]

    # ---- Auth ----
...
    Match:
      type: object
      properties:
        id: { type: integer }
        tournamentId: { type: integer }
        round: { type: integer }
        teamA: { allOf: [{ $ref: '#/components/schemas/TeamRef' }], nullable: true }
        teamB: { allOf: [{ $ref: '#/components/schemas/TeamRef' }], nullable: true }
        scheduledTime: { type: string, format: date-time, nullable: true }
        scheduledEndTime: { type: string, format: date-time, nullable: true }
        venue: { type: string, nullable: true }
        checkinOpenAt: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [scheduled, checkin_open, in_progress, completed, disputed] }
        mode: { type: string, enum: [onsite, online] }
        nextMatchId: { type: integer, nullable: true }
...
paths:
...
  /applications/{id}/approve:
    post:
      tags: [Applications]
      operationId: approveApplication
      summary: อนุมัติใบสมัคร (Organizer, Soft Filter ผ่านดุลพินิจ)
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200': { description: OK }
  /applications/{id}/reject:
    post:
      tags: [Applications]
      operationId: rejectApplication
      summary: ปฏิเสธใบสมัคร (Organizer)
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string } }
      responses:
        '200': { description: OK }
  /applications/{id}/cancel:
    post:
      tags: [Applications]
      operationId: cancelApplication
      summary: ยกเลิกการสมัครก่อนได้รับอนุมัติ (Team Leader) — FR-TR-04
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { id: {type: integer}, status: {type: string, enum: [cancelled]} }
        '409': { description: "ALREADY_DECIDED — ต้องเป็น pending เท่านั้น" }
  /applications/{id}/withdraw:
    post:
      tags: [Applications]
      operationId: withdrawApplication
      summary: ถอนตัวหลังได้รับอนุมัติ (Team Leader) — คืนช่องว่าง + แจ้ง Organizer — FR-TR-04
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200':
          description: "OK — bracketExists=true ต้องแสดงคำเตือนว่าต้องจัดสายใหม่"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  status: { type: string, enum: [withdrawn] }
                  bracketExists: { type: boolean }
  /tournaments/{id}/teams:
    get:
      tags: [Applications]
      operationId: getApprovedTeams
      security: []
      summary: รายชื่อทีมที่ได้รับอนุมัติแล้ว (สาธารณะ)
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/TeamRef' } }

  # ============ 8. BRACKETS & MATCHES ============
  /tournaments/{id}/bracket:
    post:
      tags: [Brackets & Matches]
      operationId: createBracket
      summary: สร้างสายการแข่งขัน (Organizer) — FR-MM-01
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seedingMethod]
              properties:
                seedingMethod: { type: string, enum: [random, manual] }
                manualSeeds: { type: array, items: { type: integer } }
      responses:
        '201':
          description: "สร้างสำเร็จ — บันทึกลงตาราง bracket_nodes (MySQL) ★ รอบ 5 — ถ้า bracketFormat='round_robin' ไม่สร้าง bracket_nodes เลย (nodeCount=0) และ matches ทุกแถวมี next_match_id=NULL (ข้อ 12.3)"
          content:
            application/json:
              schema:
                type: object
                properties: { matchCount: {type: integer}, bracketFormat: {type: string}, nodeCount: {type: integer} }
        '422': { description: "TEAM_COUNT_MISMATCH" }
    get:
      tags: [Brackets & Matches]
      operationId: getBracket
      security: []
      summary: "ดูสายการแข่งขัน (Organizer/สาธารณะ) — ★ รอบ 5: อ่านจาก bracket_nodes ใน MySQL (เดิมอ่านจาก MongoDB)"
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200':
          description: "OK — Round Robin คืน nodes เป็น array ว่างเสมอ (ไม่มีแนวคิด node) frontend ใช้ GET /tournaments/{id}/matches แทน"
          content:
            application/json:
              schema:
                type: object
                properties:
                  bracketFormat: { type: string }
                  nodes:
                    type: array
                    items:
                      type: object
                      properties:
                        nodeId: { type: integer }
                        bracketType: { type: string, enum: [winners, losers, grand_final] }
                        round: { type: integer, nullable: true }
                        matchNumber: { type: integer }
                        teamA: { allOf: [{ $ref: '#/components/schemas/TeamRef' }], nullable: true }
                        teamB: { allOf: [{ $ref: '#/components/schemas/TeamRef' }], nullable: true }
                        matchId: { type: integer, nullable: true }
                        matchStatus: { type: string, nullable: true }
                        advancesToNodeId:
                          type: integer
                          nullable: true
                          description: "คำนวณตอน query จาก matches.next_match_id ไม่ได้เก็บเป็นคอลัมน์จริงใน bracket_nodes (เลือกทาง B — MySQL เป็น source of truth ตัวเดียว)"
...
  /matches/{id}/schedule:
    patch:
      tags: [Brackets & Matches]
      operationId: scheduleMatch
      summary: จัดวันเวลา/สนามให้แมตช์ (Organizer) — ตรวจทับซ้อน FR-MM-02
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduledTime, scheduledEndTime, venue]
              properties:
                scheduledTime: { type: string, format: date-time }
                scheduledEndTime: { type: string, format: date-time }
                venue: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Match' }
        '409': { description: "SCHEDULE_CONFLICT — มากับ conflictingMatchId" }
...
  /matches/{id}/start:
    post:
      tags: [Brackets & Matches]
      operationId: startMatch
      summary: เริ่มการแข่งขัน (Referee)
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { id: {type: integer}, status: {type: string, enum: [in_progress]} }
        '409': { description: "INSUFFICIENT_CHECKINS" }
  /matches/{id}/checkin-qr:
    get:
      tags: [Brackets & Matches]
      operationId: getCheckinQR
      summary: QR Code สำหรับเช็คอินแบบ on-site (Organizer/Referee)
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { qrPayload: {type: string}, expiresAt: {type: string, format: date-time} }
...
  /matches/{id}/checkins/{cid}/verify:
    post:
      tags: [Brackets & Matches]
      operationId: verifyCheckin
      summary: ตรวจสอบเอกสารเช็คอินแบบ online (Referee) — FR-PV-04
      parameters:
        - { $ref: '#/components/parameters/IdParam' }
        - { name: cid, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }
  /matches/{id}/checkins/{cid}/reject:
    post:
      tags: [Brackets & Matches]
      operationId: rejectCheckin
      summary: ปฏิเสธการเช็คอิน (Referee)
      parameters:
        - { $ref: '#/components/parameters/IdParam' }
        - { name: cid, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string } }
      responses:
        '200': { description: OK }
...